Move package infos to ProductImageTests as a common denominator to fix method accessibility
