+++ tests/Microsoft.DotNet.Docker.Tests/RuntimeDepsImageTests.cs
@@ -2,11 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text.Json;
 using System.Text.Json.Nodes;
 using System.Threading.Tasks;
 using Xunit;
@@ -109,77 +105,6 @@ namespace Microsoft.DotNet.Docker.Tests
             Assert.NotEmpty(GetOSReleaseInfo(imageData, ImageRepo, DockerHelper));
         }
 
-        [LinuxImageTheory]
-        [MemberData(nameof(GetImageData))]
-        public void VerifyInstalledPackages(ProductImageData imageData)
-        {
-            VerifyInstalledPackagesBase(imageData, ImageRepo, DockerHelper, OutputHelper);
-        }
-
-        /// <summary>
-        /// Verifies that the packages installed are correct and scannable by security tools.
-        /// </summary>
-        internal static void VerifyInstalledPackagesBase(
-            ProductImageData imageData,
-            DotNetImageRepo imageRepo,
-            DockerHelper dockerHelper,
-            ITestOutputHelper outputHelper,
-            IEnumerable<string> extraExcludePaths = null)
-        {
-            IEnumerable<string> expectedPackages = GetExpectedPackages(imageData, imageRepo);
-            IEnumerable<string> actualPackages = GetInstalledPackages(imageData, imageRepo, dockerHelper, extraExcludePaths);
-
-            ComparePackages(expectedPackages, actualPackages, imageData.IsDistroless, outputHelper);
-        }
-
-        internal static void ComparePackages(
-            IEnumerable<string> expectedPackages,
-            IEnumerable<string> actualPackages,
-            bool isDistroless,
-            ITestOutputHelper outputHelper)
-        {
-            outputHelper.WriteLine($"Expected Packages: [ {string.Join(", ", expectedPackages)} ]");
-
-            // Verify we only include strictly necessary packages in distroless images
-            if (isDistroless)
-            {
-                outputHelper.WriteLine($"Actual Packages: [ {string.Join(", ", actualPackages)} ]");
-                Assert.Equal(expectedPackages, actualPackages);
-                return;
-            }
-
-            // Verify satisfy .NET dependencies on non-distroless images.
-            // There will be additional packages from the distro.
-            IEnumerable<string> missingPackages = expectedPackages.Except(actualPackages);
-            if (missingPackages.Any())
-            {
-                outputHelper.WriteLine($"Missing packages: [ {string.Join(", ", missingPackages)} ]");
-            }
-
-            Assert.Empty(missingPackages);
-        }
-
-
-        internal static IEnumerable<string> GetInstalledPackages(
-            ProductImageData imageData,
-            DotNetImageRepo imageRepo,
-            DockerHelper dockerHelper,
-            IEnumerable<string> extraExcludePaths = null)
-        {
-            JsonNode output = GetSyftOutput("package-info", imageData, imageRepo, dockerHelper, extraExcludePaths);
-            return ((JsonArray)output["artifacts"])
-                .Select(artifact => artifact["name"]?.ToString())
-                // Syft can sometimes detect duplicates of packages if they have a distro-specific version suffix. Syft
-                // will report the distro package version and the binary package version as separate.
-                //
-                // openssl                      1.1.1k                binary
-                // openssl                      1.1.1k-29.cm2         rpm
-                //                              ^ Mariner specific version
-                //
-                // So we need to remove duplicates.
-                .Distinct();
-        }
-
         private static string GetOSReleaseInfo(
             ProductImageData imageData,
             DotNetImageRepo imageRepo,
@@ -190,213 +115,15 @@ namespace Microsoft.DotNet.Docker.Tests
             return (string)distro["version"];
         }
 
-        private static JsonNode GetSyftOutput(
-            string syftContainerName,
-            ProductImageData imageData,
-            DotNetImageRepo imageRepo,
-            DockerHelper dockerHelper,
-            IEnumerable<string> extraExcludePaths = null)
+
+        [LinuxImageTheory]
+        [MemberData(nameof(GetImageData))]
+        public void VerifyInstalledPackages(ProductImageData imageData)
         {
-            string syftImage = $"{Config.GetVariableValue("syft|repo")}:{Config.GetVariableValue("syft|tag")}";
-            dockerHelper.Pull(syftImage);
-
-            string imageToInspect = imageData.GetImage(imageRepo, dockerHelper);
-
-            string outputContainerFilePath = "/artifacts/output.json";
-            string tempDir = null;
-            string outputContents = null;
-
-            // Ignore the dotnet folder, or else syft will report all the packages in the .NET Runtime. We only care
-            // about the packages from the linux distro for this test.
-            extraExcludePaths ??= [];
-            extraExcludePaths = extraExcludePaths.Append("/usr/share/dotnet");
-            IEnumerable<string> excludeArgs = extraExcludePaths.Select(path => $"--exclude {path}");
-
-            string[] args = [
-                "scan",
-                $"docker:{imageToInspect}",
-                $"-o json={outputContainerFilePath}",
-                ..excludeArgs
-            ];
-
-            try
-            {
-                dockerHelper.Run(
-                    syftImage,
-                    syftContainerName,
-                    string.Join(' ', args),
-                    skipAutoCleanup: true,
-                    useMountedDockerSocket: true);
-
-                tempDir = Directory.CreateDirectory(
-                    Path.Combine(Path.GetTempPath(), Path.GetRandomFileName())).FullName;
-
-                dockerHelper.Copy($"{syftContainerName}:{outputContainerFilePath}", tempDir);
-
-                string outputLocalFilePath = Path.Join(tempDir, Path.GetFileName(outputContainerFilePath));
-                outputContents = File.ReadAllText(outputLocalFilePath);
-            }
-            finally
-            {
-                if (!string.IsNullOrEmpty(tempDir))
-                {
-                    Directory.Delete(tempDir, true);
-                }
-                dockerHelper.DeleteContainer(syftContainerName);
-            }
-
-            return JsonNode.Parse(outputContents)
-                    ?? throw new JsonException($"Unable to parse the output as JSON:{Environment.NewLine}{outputContents}");
+            VerifyInstalledPackagesBase(imageData, ImageRepo, DockerHelper, OutputHelper);
         }
-
-        internal static IEnumerable<string> GetExpectedPackages(ProductImageData imageData, DotNetImageRepo imageRepo)
-        {
-            IEnumerable<string> expectedPackages = imageData.ImageVariant.HasFlag(DotNetImageVariant.AOT)
-                ? GetAotDepsPackages(imageData)
-                : GetRuntimeDepsPackages(imageData);
-
-            if (imageData.IsDistroless)
-            {
-                expectedPackages = [..expectedPackages, ..GetDistrolessBasePackages(imageData)];
-            }
-
-            if (imageData.ImageVariant.HasFlag(DotNetImageVariant.Extra)
-                || (imageRepo == DotNetImageRepo.SDK && imageData.Version.Major != 6 && imageData.Version.Major != 7))
-            {
-                expectedPackages = [..expectedPackages, ..GetExtraPackages(imageData)];
-            }
-
-            return expectedPackages.Distinct().OrderBy(s => s);
-        }
-
-        private static IEnumerable<string> GetDistrolessBasePackages(ProductImageData imageData) => imageData switch
-            {
-                { OS: string os } when os.Contains(OS.Mariner) => new[]
-                    {
-                        "distroless-packages-minimal",
-                        "filesystem",
-                        "mariner-release",
-                        "prebuilt-ca-certificates",
-                        "tzdata"
-                    },
-                { OS: OS.JammyChiseled } => new[]
-                    {
-                        "base-files"
-                    },
-                _ => throw new NotSupportedException()
-            };
-
-        private static IEnumerable<string> GetAotDepsPackages(ProductImageData imageData) => imageData switch
-            {
-                { OS: OS.Mariner20Distroless, Version: ImageVersion version }
-                        when version.Major == 6 || version.Major == 7 => new[]
-                    {
-                        "e2fsprogs-libs",
-                        "glibc",
-                        "krb5",
-                        "libgcc",
-                        "openssl",
-                        "openssl-libs",
-                        "prebuilt-ca-certificates",
-                        "zlib"
-                    },
-                { OS: OS.Mariner20, Version: ImageVersion version }
-                        when version.Major == 6 || version.Major == 7 => new[]
-                    {
-                        "glibc",
-                        "icu",
-                        "krb5",
-                        "libgcc",
-                        "openssl-libs",
-                        "zlib"
-                    },
-                { OS: string os } when os.Contains(OS.Mariner) => new[]
-                    {
-                        "glibc",
-                        "libgcc",
-                        "openssl-libs",
-                        "zlib"
-                    },
-                { OS: string os } when os.Contains(OS.Jammy) => new[]
-                    {
-                        "ca-certificates",
-                        "libc6",
-                        "libgcc-s1",
-                        "libssl3",
-                        "zlib1g"
-                    },
-                { OS: OS.Focal } => new[]
-                    {
-                        "ca-certificates",
-                        "libc6",
-                        "libgcc-s1",
-                        "libgssapi-krb5-2",
-                        "libicu66",
-                        "libssl1.1",
-                        "zlib1g"
-                    },
-                { OS: string os } when os.Contains(OS.Alpine) => new[]
-                    {
-                        "ca-certificates-bundle",
-                        "libgcc",
-                        "libssl3",
-                        "zlib"
-                    },
-                { OS: OS.BookwormSlim } => new[]
-                    {
-                        "ca-certificates",
-                        "libc6",
-                        "libgcc-s1",
-                        "libicu72",
-                        "libssl3",
-                        "tzdata",
-                        "zlib1g"
-                    },
-                { OS: OS.BullseyeSlim } => new[]
-                    {
-                        "ca-certificates",
-                        "libc6",
-                        "libgcc-s1", // Listed as libgcc1 in the Dockerfile
-                        "libgssapi-krb5-2",
-                        "libicu67",
-                        "libssl1.1",
-                        "zlib1g"
-                    },
-                _ => throw new NotSupportedException()
-            };
-
-        private static IEnumerable<string> GetRuntimeDepsPackages(ProductImageData imageData) {
-            string libstdcppPkgName = imageData.OS.Contains(OS.Mariner) || imageData.OS.Contains(OS.Alpine)
-                ? "libstdc++"
-                : "libstdc++6";
-            return GetAotDepsPackages(imageData).Append(libstdcppPkgName);
-        }
-
-        internal static IEnumerable<string> GetExtraPackages(ProductImageData imageData) => imageData switch
-            {
-                { IsDistroless: true, OS: string os } when os.Contains(OS.Mariner) => new[]
-                    {
-                        "icu",
-                        "tzdata"
-                    },
-                { OS: OS.JammyChiseled } => new[]
-                    {
-                        "libicu70",
-                        "tzdata"
-                    },
-                { OS: string os } when os.Contains(OS.Alpine) => new[]
-                    {
-                        "icu-data-full",
-                        "icu-libs",
-                        "tzdata"
-                    },
-                _ => Array.Empty<string>()
-            };
 
         internal static string[] GetExpectedRpmPackagesInstalled(ProductImageData imageData) =>
-            new string[]
-                {
-                    $"dotnet-runtime-deps-{imageData.VersionString}"
-                };
+            [ $"dotnet-runtime-deps-{imageData.VersionString}" ];
     }
 }
